+++ routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
-use App\Http\Controllers\views\UsersController;
-use App\Http\Controllers\views\ModeratorsController;
+use App\Http\Controllers\ModeratorsController;
+use App\Http\Controllers\UsersController;
 use Illuminate\Support\Facades\Route;
 
 /*
